requirements.txt
alembic==1.8.1
APScheduler==3.9.1
cacheout==0.14.1
cloudpickle==2.2.1
cn2an==0.5.17
dateparser==1.1.4
dill==0.3.7
email_validator==2.0.0.post2
fast_bencode==1.1.3
Flask==2.3.2
Flask_Compress==1.13
Flask-Session==0.5.0
Flask_Login==0.6.2
flask_restx==1.1.0
Jinja2==3.1.2
jsonpath==0.82
lxml==4.9.1
parse==1.19.0
pikpakapi==0.1.3
pycryptodome==3.15.0
pyinstaller==5.13.0
PyJWT==2.8.0
pypushdeer==0.0.3
pyquery==1.4.3
python_dateutil==2.8.2
pytz==2022.2.1
PyVirtualDisplay==3.0
qbittorrent_api==2023.6.50
regex==2022.9.13
Requests==2.31.0
rsa==4.9
ruamel.base==1.0.0
selenium==4.12.0
setuptools==68.0.0
SQLAlchemy==1.4.42
transmission_rpc==6.0.0
undetected_chromedriver==3.1.7
urllib3==1.26.12
watchdog==2.1.9
webdriver_manager==3.8.5
Werkzeug==2.3.6
wordninja==2.0.0
zhconv==1.4.3
feapder==1.8.8
anitopy==2.1.1
plexapi==4.15.1
slack-bolt==1.18.0
passlib==1.7.4
wxPython==4.2.1

